Split large TradeAsset component file
